storelist 클래스
